⚡️ :: Long 타입 변환

+++ src/main/java/com/example/dailyboramspring/domain/series/presentation/dto/response/FeedListElement.java
@@ -17,7 +17,7 @@ public class FeedListElement {
 
     private String nickname;
 
-    private Integer like;
+    private Long like;
 
     private List<String> genre;
 }
